+++ changelog.md
@@ -15,6 +15,8 @@ Codion Change Log
 - ConnectionPoolWrapper.VALIDATE_CONNECTIONS_ON_CHECKOUT added.
 ### is.codion.common.model
 - DefaultFilterModelItems.filterIncremental() added.
+### is.codion.common.i18n
+- Messages, mnemonic bounds checking added.
 ### is.codion.dbms.h2
 - H2Database, security validation for the H2 database script injection.
 ### is.codion.swing.common.ui